Started working on backend and changed few structure (#3)
* WIP: Changed folder structure and reorganised few files

* Refactored tokens and query routes (Functionality not tested)

* Update requirements.txt file

* Transferred sessions code into seperate files

* Added github actions file

* Updated actions file

* Added health router and also updated the guthub workflows

* Updated workflows

* Removed starting of fastapi app in background

* Updated workflows

* Declarative base was wrongly imported due to refactoring

* Updated workflow to start fastapi app in background

* Updated workfflows

* Workflows were unable to kill the started fastapi app

* Removed the kill fastapi app as its not required
