+++ .github/workflows/ci.yml
@@ -1,0 +1,40 @@
+name: FastAPI CI
+
+on: [push, pull_request] 
+
+jobs:
+  test:
+    runs-on: ubuntu-latest 
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v3  
+
+      - name: Set up Python
+        uses: actions/setup-python@v4  
+        with:
+          python-version: '3.10'
+          cache: 'pip'  
+
+      - name: Install dependencies
+        run: |
+          cd backend
+          pip install -r requirements.txt
+
+      - name: Start FastAPI app
+        run: |
+          cd backend
+          nohup fastapi dev > /dev/null 2>&1 &
+          echo $! > uvicorn.pid 
+          sleep 10 
+
+      - name: Check FastAPI /health endpoint
+        run: |
+          response=$(curl -s http://127.0.0.1:8000/health)
+          echo "Response: $response"
+          if [ "$response" == '{"status":"Active"}' ]; then
+            echo "Health check passed"
+          else
+            echo "Health check failed"
+            exit 1
+          fi
