+++ backend/api/health.py
@@ -1,0 +1,9 @@
+
+from fastapi import APIRouter
+
+health_router = APIRouter()
+
+
+@health_router.get("/health")
+async def check_health():
+    return {"status": "Active"}
