+++ backend/config.py
@@ -1,0 +1,13 @@
+SECRET_KEY = "your_secret_key"
+ALGORITHM = "HS256"
+ACCESS_TOKEN_EXPIRE_MINUTES = 30
+
+# No origins are limited for now
+# All origins are allowed for now and this 
+# must be changed
+origins = [
+    "http://localhost",
+    "http://127.0.0.1:8000",
+]
+
+
