+++ backend/requirements.txt
@@ -11,14 +11,18 @@ bcrypt==4.2.0
 build==1.2.1
 cachetools==5.4.0
 certifi==2024.7.4
+cffi==1.16.0
+cfgv==3.4.0
 charset-normalizer==3.3.2
 chroma-hnswlib==0.7.6
 chromadb==0.5.5
 click==8.1.7
 colorama==0.4.6
 coloredlogs==15.0.1
+cryptography==43.0.0
 dataclasses-json==0.6.7
 Deprecated==1.2.14
+distlib==0.3.8
 dnspython==2.6.1
 email_validator==2.2.0
 exceptiongroup==1.2.2
@@ -38,12 +42,14 @@ httptools==0.6.1
 httpx==0.27.0
 huggingface-hub==0.24.5
 humanfriendly==10.0
+identify==2.6.0
 idna==3.7
 importlib_metadata==8.0.0
 importlib_resources==6.4.0
 Jinja2==3.1.4
 jsonpatch==1.33
 jsonpointer==3.0.0
+jwt==1.3.1
 kubernetes==30.1.0
 langchain==0.2.12
 langchain-chroma==0.1.2
@@ -62,6 +68,7 @@ monotonic==1.6
 mpmath==1.3.0
 multidict==6.0.5
 mypy-extensions==1.0.0
+nodeenv==1.9.1
 numpy==1.26.4
 oauthlib==3.2.2
 ollama==0.3.1
@@ -79,10 +86,12 @@ opentelemetry-util-http==0.47b0
 orjson==3.10.6
 overrides==7.7.0
 packaging==23.2
+platformdirs==4.2.2
 posthog==3.5.0
 protobuf==4.25.4
 pyasn1==0.6.0
 pyasn1_modules==0.4.0
+pycparser==2.22
 pydantic==2.8.2
 pydantic_core==2.20.1
 Pygments==2.18.0
@@ -114,6 +123,7 @@ typing-inspect==0.9.0
 typing_extensions==4.12.2
 urllib3==2.2.2
 uvicorn==0.30.5
+virtualenv==20.26.3
 watchfiles==0.22.0
 websocket-client==1.8.0
 websockets==12.0
