+++ backend/sessions.py
@@ -1,0 +1,13 @@
+from sqlalchemy import create_engine
+
+from sqlalchemy.orm import sessionmaker
+
+from backend.sqlalchemy_models import Base
+
+
+engine = create_engine("sqlite:///users.db")
+
+Base.metadata.create_all(engine)
+
+session_maker = sessionmaker(bind=engine)
+session = session_maker()
